+++ front/temporal/agent_loop/activities/create_tool_actions.ts
@@ -1,7 +1,4 @@
-import assert from "assert";
-
 import type {
-  ActionBaseParams,
   MCPActionType,
   MCPToolConfigurationType,
 } from "@app/lib/actions/mcp";
@@ -13,15 +10,14 @@ import type { AuthenticatorType } from "@app/lib/auth";
 import { Authenticator } from "@app/lib/auth";
 import type { AgentMCPAction } from "@app/lib/models/assistant/actions/mcp";
 import type { AgentMessage } from "@app/lib/models/assistant/conversation";
-import { AgentStepContentResource } from "@app/lib/resources/agent_step_content_resource";
 import { updateResourceAndPublishEvent } from "@app/temporal/agent_loop/activities/common";
+import { buildActionBaseParams } from "@app/temporal/agent_loop/lib/action_utils";
 import type {
   AgentActionsEvent,
   AgentConfigurationType,
   AgentMessageType,
   ModelId,
 } from "@app/types";
-import { isFunctionCallContent } from "@app/types/assistant/agent_message_content";
 import type { RunAgentArgs } from "@app/types/assistant/agent_run";
 import { getRunAgentData } from "@app/types/assistant/agent_run";
 
@@ -114,35 +110,15 @@ async function createActionForTool(
     step: number;
   }
 ): Promise<ActionBlob | void> {
-  // Fetch step content to derive inputs, functionCallId, and step.
-  const stepContent =
-    await AgentStepContentResource.fetchByModelId(stepContentId);
-  assert(
-    stepContent,
-    `Step content not found for stepContentId: ${stepContentId}`
-  );
-
-  if (!isFunctionCallContent(stepContent.value)) {
-    throw new Error(
-      `Expected step content to be a function call, got: ${stepContent.value.type}`
-    );
-  }
-
-  const rawInputs = JSON.parse(stepContent.value.value.arguments);
-  const { id: functionCallId } = stepContent.value.value;
-
-  const actionBaseParams: ActionBaseParams = {
+  const actionBaseParams = await buildActionBaseParams({
     agentMessageId: agentMessage.agentMessageId,
     citationsAllocated: stepContext.citationsCount,
-    functionCallId,
-    functionCallName: actionConfiguration.name,
-    generatedFiles: [],
     mcpServerConfigurationId: actionConfiguration.id.toString(),
-    params: rawInputs,
     step,
-  };
+    stepContentId,
+  });
 
-  const validateToolInputsResult = validateToolInputs(rawInputs);
+  const validateToolInputsResult = validateToolInputs(actionBaseParams.params);
   if (validateToolInputsResult.isErr()) {
     return updateResourceAndPublishEvent(
       {
