+++ front/temporal/agent_loop/activities/run_tool.ts
@@ -1,21 +1,17 @@
 import assert from "assert";
 
-import type {
-  ActionBaseParams,
-  MCPToolConfigurationType,
-} from "@app/lib/actions/mcp";
+import type { MCPToolConfigurationType } from "@app/lib/actions/mcp";
 import { MCPActionType } from "@app/lib/actions/mcp";
 import { runToolWithStreaming } from "@app/lib/actions/mcp";
 import type { StepContext } from "@app/lib/actions/types";
 import type { AuthenticatorType } from "@app/lib/auth";
 import { Authenticator } from "@app/lib/auth";
 import { AgentMCPAction } from "@app/lib/models/assistant/actions/mcp";
-import { AgentStepContentResource } from "@app/lib/resources/agent_step_content_resource";
 import { updateResourceAndPublishEvent } from "@app/temporal/agent_loop/activities/common";
+import { buildActionBaseParams } from "@app/temporal/agent_loop/lib/action_utils";
 import { sliceConversationForAgentMessage } from "@app/temporal/agent_loop/lib/loop_utils";
 import type { ModelId } from "@app/types";
 import { assertNever } from "@app/types";
-import { isFunctionCallContent } from "@app/types/assistant/agent_message_content";
 import type { RunAgentArgs } from "@app/types/assistant/agent_run";
 import { getRunAgentData } from "@app/types/assistant/agent_run";
 
@@ -69,28 +65,13 @@ export async function runToolActivity(
   const action = await AgentMCPAction.findByPk(actionId);
   assert(action, "Action not found");
 
-  const stepContent = await AgentStepContentResource.fetchByModelId(
-    action.stepContentId
-  );
-  assert(stepContent, "Step content not found");
-
-  const stepContentJSON = stepContent.toJSON();
-  const { value: functionCallData } = stepContentJSON;
-  assert(
-    isFunctionCallContent(functionCallData),
-    "Step content is not a function call"
-  );
-
-  const actionBaseParams: ActionBaseParams = {
+  const actionBaseParams = await buildActionBaseParams({
     agentMessageId: action.agentMessageId,
     citationsAllocated: action.citationsAllocated,
-    functionCallId: functionCallData.value.id,
-    functionCallName: actionConfiguration.name,
-    generatedFiles: [],
     mcpServerConfigurationId: action.mcpServerConfigurationId,
-    params: JSON.parse(functionCallData.value.arguments),
     step,
-  };
+    stepContentId: action.stepContentId,
+  });
 
   const eventStream = runToolWithStreaming(auth, actionConfiguration, {
     action,
